Extract ProfilesSelection from BrowseContainer
