+++ src/Containers/ProfilesSelection.js
@@ -2,7 +2,15 @@
 import { Header, Profiles } from 'components';
 import routes from 'routes';
 
-const ProfilesSelection = ({ user, setProfile }) => {
+const ProfilesSelection = ({ user, setProfile, setLoading }) => {
+  const handleProfileClick = () => {
+    setProfile({
+      displayName: user.displayName,
+      photoURL: user.photoURL,
+    });
+    setLoading((loading) => !loading);
+  };
+
   return (
     <>
       <Header isNotBrowse isProfiles>
@@ -17,14 +25,7 @@ const ProfilesSelection = ({ user, setProfile }) => {
       <Profiles>
         <Profiles.Title>Who&apos;s watching?</Profiles.Title>
         <Profiles.List>
-          <Profiles.Item
-            onClick={() =>
-              setProfile({
-                displayName: user.displayName,
-                photoURL: user.photoURL,
-              })
-            }
-          >
+          <Profiles.Item onClick={() => handleProfileClick()}>
             <Profiles.Avatar src={1} />
             <Profiles.Name>{user.displayName}</Profiles.Name>
           </Profiles.Item>
